+++ pages/api/webhook.js
@@ -1,5 +1,5 @@
 import {initMongoose} from "../../lib/mongoose";
-import Order from "../../modes/Order";
+import Order from "../../modelss/Order";
 import {buffer} from 'micro';
 const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY);
